split up mountains file, made mountains class more flexible
