+++ core/support/mountain.py
@@ -1,8 +1,10 @@
-from dataclasses import dataclass, fields
+from dataclasses import dataclass, fields, field
 from typing import Self, Optional
 from datetime import datetime
 
 from core.support.states import State, Region
+from core.support.trail import Trail
+from core.support.lift import Lift
 
 
 @dataclass
@@ -18,12 +20,12 @@ class Mountain:
     state: State
     direction: str
     season_passes: list
-    trail_count: int
-    lift_count: int
-    vertical: int
-    difficulty: float
-    beginner_friendliness: float
+    vertical: Optional[int] = None
+    difficulty: Optional[float] = None
+    beginner_friendliness: Optional[float] = None
     last_updated: Optional[datetime] = datetime.now()
+    trails: Optional[dict] = field(default_factory=dict)
+    lifts: Optional[dict] = field(default_factory=dict)
 
     def region(self) -> Region:
         """
@@ -46,88 +48,33 @@ class Mountain:
             return 90
         raise ValueError(f"Invalid direction value: {self.direction}")
 
+    def trail_count(self) -> int:
+        """
+        Returns the number of trails associated with the Mountain
+        """
+        return len(self.trails)
+
+    def lift_count(self) -> int:
+        """
+        Returns the number of lifts associated with the Mountain
+        """
+        return len(self.lifts)
+
+    def add_trail(self, trail: Trail) -> None:
+        """
+        Inserts a new trail into trails dict
+        """
+        self.trails[trail.id] = Trail
+
+    def add_lift(self, lift: Lift) -> None:
+        """
+        Inserts a new trail into lifts dict
+        """
+        self.lifts[lift.id] = Lift
+
     def from_db(id: str) -> Self:
         """
         Gets mountain data from database and returns a Mountain object
-        """
-        return "TODO"
-
-    def to_db(self) -> None:
-        # check that all fields have been populated before saving
-        missing_fields = [f.name for f in fields(self) if getattr(self, f.name) is None]
-        if len(missing_fields) > 0:
-            raise ValueError(f"The following fields are missing: {missing_fields}")
-        """
-        Updates DB record with the values in the dataclass
-        """
-        return "TODO"
-
-
-@dataclass
-class Trail:
-    """
-    Trail dataclass that contains all information about a specific trail.
-    An existing trail can be loaded from the DB with from_db, and a new
-    or updated trail can be saved back to the DB with to_db.
-    """
-
-    id: str
-    mountain_id: int
-    geometry: str
-    name: str
-    official_rating: str
-    gladed: bool
-    area: bool
-    ungroomed: bool
-    park: bool
-    length: Optional[float] = None
-    vertical: Optional[float] = None
-    difficulty: Optional[float] = None
-    max_slope: Optional[float] = None
-    average_slope: Optional[float] = None
-
-    def from_db(id: str) -> Self:
-        """
-        Gets trail data from database and returns a Trail object
-        """
-        return "TODO"
-
-    def to_db(self) -> None:
-        # check that all fields have been populated before saving
-        missing_fields = [f.name for f in fields(self) if getattr(self, f.name) is None]
-        if len(missing_fields) > 0:
-            raise ValueError(f"The following fields are missing: {missing_fields}")
-        """
-        Updates DB record with the values in the dataclass
-        """
-        return "TODO"
-
-
-@dataclass
-class Lift:
-    """
-    Lift dataclass that contains all information about a specific lift.
-    An existing lift can be loaded from the DB with from_db, and a new
-    or updated lift can be saved back to the DB with to_db.
-    """
-
-    id: str
-    mountain_id: int
-    geometry: str
-    name: str
-    lift_type: str
-    occupancy: int
-    capacity: int
-    detatchable: bool
-    bubble: bool
-    heating: bool
-    length: Optional[float] = None
-    vertical: Optional[float] = None
-    average_slope: Optional[float] = None
-
-    def from_db(id: str) -> Self:
-        """
-        Gets lift data from database and returns a Lift object
         """
         return "TODO"
